refactor: 프로젝트 멤버 조회 응답 형태 수정
